Added sign up pages for customers and organizations
