+++ nextjs-dashboard/app/signup/organization/layout.tsx
@@ -1,0 +1,13 @@
+import '@/app/ui/global.css';
+
+export default function RootLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return (
+    <html lang="en">
+      <body>{children}</body>
+    </html>
+  );
+}
