+++ nextjs-dashboard/app/signup/signup.tsx
@@ -6,6 +6,8 @@ import {
   ExclamationCircleIcon,
   PencilSquareIcon,
   PhoneIcon,
+  HomeIcon,
+  GlobeAltIcon,
 } from '@heroicons/react/24/outline';
 import { ArrowRightIcon } from '@heroicons/react/20/solid';
 import { Button } from '../ui/button';
@@ -13,159 +15,33 @@ import Link from 'next/link';
 import { useState } from 'react';
 import { useRouter } from 'next/navigation';
 
-export default function SignUpForm() {
-  const [email, setEmail] = useState('');
-  const [password, setPassword] = useState('');
-  const [first, setFirst] = useState('');
-  const [last, setLast] = useState('');
-  const [phone, setPhone] = useState('');
-  const [formError, setFormError] = useState('');
+
+export default function UserOrOrganization() {
   const router = useRouter();
+  const handleCustomer = () =>{
+    router.push('/signup/customer');
+  }
 
-  const handleSubmit = (e: React.FormEvent<HTMLFormElement>) => {
-    e.preventDefault();
-
-    // Simple validation
-    if (!email || !password || !first || !last || !phone) {
-      setFormError('Please fill in all fields.');
-      return;
-    }
-
-    // Additional validation logic can go here (e.g., email format, password length)
-
-    // If validation passes, redirect to the dashboard
-    setFormError(''); // Clear any previous errors
-    router.push('/dashboard');
-  };
+  const handleOrg = () =>{
+    router.push('/signup/organization');
+  }
   return (
-    <form className="space-y-3" onSubmit={handleSubmit}>
-      <div className="flex-1 rounded-lg bg-gray-50 px-6 pb-4 pt-8">
-        <h1 className={`${lusitana.className} mb-3 text-2xl`}>
-          Please fill in the following fields to sign up.
-        </h1>
-        <div className="w-full">
-          <div>
-            <label
-              className="mb-3 mt-5 block text-xs font-medium text-gray-900"
-              htmlFor="first_name"
-            >
-              First Name
-            </label>
-            <div className="relative">
-              <input
-                className="peer block w-full rounded-md border border-gray-200 py-[9px] pl-10 text-sm outline-2 placeholder:text-gray-500"
-                id="first_name"
-                type="first_name"
-                name="first_name"
-                placeholder="Enter your First Name"
-                required
-                value={first}
-                onChange={(e) => setFirst(e.target.value)}
-              />
-              <PencilSquareIcon className="pointer-events-none absolute left-3 top-1/2 h-[18px] w-[18px] -translate-y-1/2 text-gray-500 peer-focus:text-gray-900" />
-            </div>
-          </div>
-          <div>
-            <label
-              className="mb-3 mt-5 block text-xs font-medium text-gray-900"
-              htmlFor="last_name"
-            >
-              Last Name
-            </label>
-            <div className="relative">
-              <input
-                className="peer block w-full rounded-md border border-gray-200 py-[9px] pl-10 text-sm outline-2 placeholder:text-gray-500"
-                id="last_name"
-                type="last_name"
-                name="last_name"
-                placeholder="Enter your Last Name"
-                required
-                value={last}
-                onChange={(e) => setLast(e.target.value)}
-              />
-              <PencilSquareIcon className="pointer-events-none absolute left-3 top-1/2 h-[18px] w-[18px] -translate-y-1/2 text-gray-500 peer-focus:text-gray-900" />
-            </div>
-          </div>
-          <div>
-            <label
-              className="mb-3 mt-5 block text-xs font-medium text-gray-900"
-              htmlFor="email"
-            >
-              Email
-            </label>
-            <div className="relative">
-              <input
-                className="peer block w-full rounded-md border border-gray-200 py-[9px] pl-10 text-sm outline-2 placeholder:text-gray-500"
-                id="email"
-                type="email"
-                name="email"
-                placeholder="Enter your email address"
-                required
-                value={email}
-                onChange={(e) => setEmail(e.target.value)}
-              />
-              <AtSymbolIcon className="pointer-events-none absolute left-3 top-1/2 h-[18px] w-[18px] -translate-y-1/2 text-gray-500 peer-focus:text-gray-900" />
-            </div>
-          </div>
-          <div className="mt-4">
-            <label
-              className="mb-3 mt-5 block text-xs font-medium text-gray-900"
-              htmlFor="password"
-            >
-              Password
-            </label>
-            <div className="relative">
-              <input
-                className="peer block w-full rounded-md border border-gray-200 py-[9px] pl-10 text-sm outline-2 placeholder:text-gray-500"
-                id="password"
-                type="password"
-                name="password"
-                placeholder="Enter password"
-                required
-                minLength={6}
-                value={password}
-                onChange={(e) => setPassword(e.target.value)}
-              />
-              <KeyIcon className="pointer-events-none absolute left-3 top-1/2 h-[18px] w-[18px] -translate-y-1/2 text-gray-500 peer-focus:text-gray-900" />
-            </div>
-          </div>
-          <div className="mt-4">
-            <label
-              className="mb-3 mt-5 block text-xs font-medium text-gray-900"
-              htmlFor="phone"
-            >
-              Phone Number
-            </label>
-            <div className="relative">
-              <input
-                className="peer block w-full rounded-md border border-gray-200 py-[9px] pl-10 text-sm outline-2 placeholder:text-gray-500"
-                id="phone"
-                type="phone"
-                name="phone"
-                placeholder="Enter your phone number"
-                required
-                minLength={6}
-                value={phone}
-                onChange={(e) => setPhone(e.target.value)}
-              />
-              <PhoneIcon className="pointer-events-none absolute left-3 top-1/2 h-[18px] w-[18px] -translate-y-1/2 text-gray-500 peer-focus:text-gray-900" />
-            </div>
-          </div>
-        </div>
-
-        {formError && (
-          <div className="mt-2 text-sm text-red-600">
-            <ExclamationCircleIcon className="inline h-5 w-5 mr-1" />
-            {formError}
-          </div>
-        )}
-
+    <div className="flex-1 rounded-lg bg-gray-50 px-6 pb-4 pt-8">
+      <h1 className={`${lusitana.className} mb-3 text-2xl`}>
+        Select one of the following options.
+      </h1>
+      <div className="w-full">
         <div>
-          <Button type="submit" className="mt-4 w-full">
-            Sign Up <ArrowRightIcon className="ml-auto h-5 w-5 text-gray-50" />
+          <Button className="mt-4 w-full" onClick={handleCustomer}>
+            Customer <ArrowRightIcon className="ml-auto h-5 w-5 text-gray-50" />
+          </Button>
+        </div>
+        <div>
+          <Button type="submit" className="mt-4 w-full" onClick={handleOrg}>
+            Organization <ArrowRightIcon className="ml-auto h-5 w-5 text-gray-50" />
           </Button>
         </div>
       </div>
-    </form>
-  );
+    </div>
+  )
 }
